# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/mindmaps": {
      "get": {
        "operationId": "MindmapsController_findMine",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MindmapDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "mindmaps"
        ]
      },
      "post": {
        "operationId": "MindmapsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMindmapDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MindmapDto"
                }
              }
            }
          },
          "400": {
            "description": "Body failed validation"
          }
        },
        "tags": [
          "mindmaps"
        ]
      }
    },
    "/api/mindmaps/{id}": {
      "get": {
        "operationId": "MindmapsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MindmapDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "mindmaps"
        ]
      },
      "patch": {
        "operationId": "MindmapsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMindmapDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MindmapDto"
                }
              }
            }
          },
          "400": {
            "description": "Body failed validation, or the patch would leave the mindmap with a graph that is not a tree"
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "mindmaps"
        ]
      },
      "delete": {
        "operationId": "MindmapsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "mindmaps"
        ]
      }
    },
...
  },
  "info": {
    "title": "ThinkClear API",
    "description": "ThinkClear app API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token"
      }
    },
    "schemas": {
...
      "MindmapNodeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "root",
            "description": "Unique within the mindmap. The node with id \"root\" is the map's root and cannot be removed."
          },
          "title": {
            "type": "string",
            "example": "Backend",
            "minLength": 1,
            "maxLength": 200
          },
          "x": {
            "type": "number",
            "example": 0,
            "description": "Canvas position. The editor lays the tree out from scratch on load, so this is a hint about sibling order rather than a fixed coordinate."
          },
          "y": {
            "type": "number",
            "example": 0
          },
          "note": {
            "type": "string",
            "example": "Ship the **API** first, then the canvas.",
            "maxLength": 5000,
            "description": "The topic's note, as markdown source. Omitted when the topic has no note — an empty string is not stored."
          }
        },
        "required": [
          "id",
          "title",
          "x",
          "y"
        ]
      },
      "MindmapEdgeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique within the mindmap"
          },
          "source": {
            "type": "string",
            "description": "Node id the edge starts from"
          },
          "target": {
            "type": "string",
            "description": "Node id the edge points to. Direction is presentation only — the editor re-points edges parent to child when it lays the tree out."
          }
        },
        "required": [
          "id",
          "source",
          "target"
        ]
      },
      "MindmapDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "ownerId": {
            "type": "string"
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MindmapNodeDto"
            }
          },
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MindmapEdgeDto"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "_id",
          "title",
          "ownerId",
          "nodes",
          "edges",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateMindmapDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "My first mindmap"
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateMindmapDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Renamed mindmap",
            "minLength": 1,
            "maxLength": 200
          },
          "nodes": {
            "maxItems": 500,
            "description": "Replaces every node. Ids must be unique and one of them must be \"root\".",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MindmapNodeDto"
            }
          },
          "edges": {
            "maxItems": 1000,
            "description": "Replaces every edge. Ids must be unique, both endpoints must be nodes of this mindmap, and the result must stay a tree: no self-connections, no node pair connected twice, and no loops. Disconnected branches are allowed.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MindmapEdgeDto"
            }
          }
        }
      },
...
    }
  }
}